wrap JS into window.onload event
this was the problem: on Mac OS Safari browser, (inexplicably) the JS seemed to execute before the :root style definitions had loaded. so, when we tried to fetch custom-properties from the stylesheet (eg. the colour value for '--foliage'), Safari was unable to fetch values, and it ended up drawing nothing. the problem perhaps lay somewhere else, but, in the absence of any more insight, wrapping the JS into a window.onload event seemed to fix it.